// store/auth/auth.type.ts
// auth.type.ts

interface Gym {
  id: number;
  gym: string;
  benefits: string | null;
  logo: string;
  linkPrivacy: string;
  linkTerms: string;
  linkSite: string;
  themeColor: string | null;
  appIos: string;
  appAndroid: string;
  status: number;
  isDeleted: number;
  createdAt: string;
  updatedAt: string;
  linkSubscription: string | null;
}

interface Branch {
  id: number;
  gymId: number;
  branch: string;
  typeBranch: string;
  slug: string;
  status: number;
  secretId: number;
  vimeoUser: string | null;
  vimeoId: string | null;
  vimeoSecret: string | null;
  vimeoToken: string | null;
  country: string;
  countryCode: string;
  codeZip: string;
  state: string;
  stateCode: string;
  population: string;
  adress: string;
  email: string;
  web: string;
  phone: string;
  communication: number;
  description: string | null;
  pathFile: string;
  subscription: string;
  externalSource: string | null;
  isChatExternal: number;
  isChat: number;
  setChatbot: string | null;
  createdAt: string;
  updatedAt: string;
  stripe: number;
  nutrition: number;
  videoClass: number;
  appPusher: number;
  lesson: number;
  pdf: string | null;
  rulePenalty: number;
  ruleMax: string;
  firstColor: string;
  secoundColor: string;
  logo: string;
  videoTutorial: string;
  chatbotId: number;
  blockNutrition: number;
  blockTraining: number;
  stripeCustomerId: string;
  usesSubscription: number;
  usesBrand: number;
  customerToGymSubscription: number;
  timezone: string;
  hourDifference: number;
  appIos: string;
  appAndroid: string;
  largeLogo: string;
  contractSigning: number;
  contractSigningImg: string;
  signatureDate: string | null;
  contractId: number;
  camerasModule: number;
  lessonModule: number;
  assistedModule: number;
  logoColorModule: number;
  videosModule: number;
  postModule: number;
  instagramModule: number;
  accessPoint: string | null;
  accessModule: number;
  accessModuleType: number;
  paymentCardModule: number;
  trialPeriodEndDate: string;
  subscriptionBlocking: number;
  ticketModule: number;
  complianceSchedule: number;
  setPost: string | null;
  store: number;
  inAppStore: number;
  storeModule: number;
  trafficLights: string;
  directDebitModule: number;
  slots: number;
  catchmentModule: number;
  recoveryModule: number;
  accountingModule: number;
  gtiCode: string | null;
  gtiUser: string | null;
  gtiPassword: string | null;
  gtiTerminal: string | null;
  additionalVideosModule: number;
  cancellationAssisteds: number;
  cancellationLessons: number;
  cancellationSlots: number;
  contractUrl: string | null;
  gateway: string;
  publicKeyPayment: string;
  privateKeyPayment: string;
  prodIntegrity: string | null;
  taxAuthorities: string | null;
  timeAccess: number;
  courseModule: number;
  iva: number;
  commentTicket: string;
  taxId: string;
  fitmewiseAccessImage: string | null;
  multiCenterQuotas: number;
  contractUser: string | null;
  clausesUser: string | null;
  daysAccess: string | null;
  assistedAnticipationDays: string | null;
  assistedTimeAccess: string | null;
  courseAnticipationDays: string | null;
  courseTimeAccess: number;
  weightUnit: string;
  measurementUnit: string;
  gym: Gym;
  contract: string;
  hasGymLink: boolean;
  hasTrainerLink: boolean;
}

export interface User {
  id: number;
  firstName: string;
  lastName: string;
  documentId: string;
  email: string;
}